Minor schema updates

+++ src/server/schema/coinbase.schema.ts
@@ -12,12 +12,12 @@ export type CreateCodeOrder = z.infer<typeof createCodeOrderSchema>;
 export const createCodeOrderSchema = z.discriminatedUnion('type', [
   z.object({
     type: z.literal('Buzz'),
-    buzzAmount: z.number().min(1000),
+    buzzAmount: z.number().int().min(1000),
   }),
   z.object({
     type: z.literal('Membership'),
     tier: z.enum(['bronze', 'silver', 'gold']),
-    months: z.number().min(1).max(12),
+    months: z.number().int().min(1).max(12),
   }),
 ]);
